+++ spec/features/users_can_see_environment_variables_spec.rb
@@ -14,7 +14,9 @@ feature "Users can see environment variables" do
 
     stub_call_to_aws_for_environment_variables(
       account_id: infrastructure.account_id,
-      request_path: "/test-app/test-service/staging/",
+      infrastructure_name: infrastructure.identifier,
+      service_name: "test-service",
+      environment_name: "staging",
       environment_variables: fake_environment_variables
     )