+++ spec/features/users_can_see_infrastructure_variables_spec.rb
@@ -11,9 +11,9 @@ feature "Users can see infrastructure variables" do
       fake_environment_variable
     ])
 
-    stub_call_to_aws_for_environment_variables(
-      account_id: infrastructure.account_id,
-      request_path: "/dalmatian-variables/infrastructures/test-app/staging/",
+    stub_call_to_aws_for_infrastructure_variables(
+      service_name: "test-app",
+      environment_name: "staging",
       environment_variables: fake_environment_variables
     )