+++ spec/support/aws_api_helpers.rb
@@ -1,4 +1,6 @@
 module AwsApiHelpers
+  CORE_AWS_ACCOUNT_ID = "0011122233344".freeze
+
   def create_aws_environment_variable(name:, value:)
     Aws::SSM::Types::Parameter.new(
       name: name,
@@ -14,12 +16,36 @@ module AwsApiHelpers
   end
 
   def stub_call_to_aws_for_environment_variables(
+    aws_ssm_client: nil,
     account_id:,
-    aws_ssm_client_double: nil,
-    request_path:,
+    infrastructure_name:,
+    service_name:,
+    environment_name:,
     environment_variables:
   )
-    aws_ssm_client = aws_ssm_client_double || stub_aws_ssm_client(account_id: account_id)
+    aws_ssm_client ||= stub_aws_ssm_client(account_id: account_id)
+    request_path = "/#{infrastructure_name}/#{service_name}/#{environment_name}/"
+
+    allow(aws_ssm_client)
+      .to receive(:get_parameters_by_path)
+      .with(
+        path: request_path,
+        with_decryption: true,
+        recursive: false
+      ).and_return(environment_variables)
+  end
+
+  def stub_call_to_aws_for_infrastructure_variables(
+    aws_ssm_client: nil,
+    service_name:,
+    environment_name:,
+    environment_variables:
+  )
+    aws_ssm_client ||= stub_aws_ssm_client(
+      aws_sts_client: stub_main_aws_sts_client,
+      account_id: CORE_AWS_ACCOUNT_ID
+    )
+    request_path = "/dalmatian-variables/infrastructures/#{service_name}/#{environment_name}/"
 
     allow(aws_ssm_client)
       .to receive(:get_parameters_by_path)
@@ -31,7 +57,7 @@ module AwsApiHelpers
   end
 
   def stub_call_to_aws_to_update_environment_variables(
-    aws_ssm_client_double: nil,
+    aws_ssm_client: nil,
     account_id:,
     infrastructure_identifier:,
     service_name:,
@@ -39,7 +65,7 @@ module AwsApiHelpers
     variable_name:,
     variable_value:
   )
-    aws_ssm_client = aws_ssm_client_double || stub_aws_ssm_client(account_id: account_id)
+    aws_ssm_client ||= stub_aws_ssm_client(account_id: account_id)
 
     path = "/#{infrastructure_identifier}/#{service_name}/#{environment_name}/"
     name_with_path = "#{path}#{variable_name}"
@@ -59,14 +85,14 @@ module AwsApiHelpers
   end
 
   def stub_call_to_aws_to_update_infrastructure_variables(
-    aws_ssm_client_double: nil,
+    aws_ssm_client: nil,
     account_id:,
     infrastructure_identifier:,
     environment_name:,
     variable_name:,
     variable_value:
   )
-    aws_ssm_client = aws_ssm_client_double || stub_aws_ssm_client(account_id: account_id)
+    aws_ssm_client ||= stub_aws_ssm_client(account_id: account_id)
 
     path = "/dalmatian-variables/infrastructures/#{infrastructure_identifier}/#{environment_name}/"
     name_with_path = "#{path}#{variable_name}"
@@ -86,21 +112,39 @@ module AwsApiHelpers
   end
 
   def stub_call_to_aws_to_delete_environment_variable(
-    aws_ssm_client_double: nil,
+    aws_ssm_client: nil,
     account_id:,
-    request_path:,
-    name:
+    infrastructure_name:,
+    service_name:,
+    environment_name:,
+    variable_name:
   )
 
-    aws_ssm_client = aws_ssm_client_double || stub_aws_ssm_client(account_id: account_id)
+    aws_ssm_client ||= stub_aws_ssm_client(account_id: account_id)
+    full_name = "#{infrastructure_name}/#{service_name}/#{environment_name}/#{variable_name}"
 
-    full_name = "#{request_path}/#{name}"
     allow(aws_ssm_client)
       .to receive(:delete_parameter)
       .with(name: full_name)
   end
 
-  def stub_aws_ssm_client(account_id:)
+  def stub_call_to_aws_to_delete_infrastructure_variable(
+    aws_ssm_client: nil,
+    account_id:,
+    service_name:,
+    environment_name:,
+    variable_name:
+  )
+
+    aws_ssm_client ||= stub_aws_ssm_client(account_id: account_id)
+    full_name = "/dalmatian-variables/infrastructures/#{service_name}/#{environment_name}/#{variable_name}"
+
+    allow(aws_ssm_client)
+      .to receive(:delete_parameter)
+      .with(name: full_name)
+  end
+
+  def stub_aws_ssm_client(aws_sts_client: stub_aws_sts_client, account_id:)
     credentials = instance_double(Aws::AssumeRoleCredentials)
     allow(Aws::AssumeRoleCredentials).to receive(:new).with(
       client: aws_sts_client,
@@ -108,18 +152,27 @@ module AwsApiHelpers
       role_session_name: "role_session_name"
     ).and_return(credentials)
 
-    aws_ssm_client_double = instance_double(Aws::SSM::Client)
+    aws_ssm_client = instance_double(Aws::SSM::Client)
     allow(Aws::SSM::Client)
       .to receive(:new)
       .with(credentials: credentials)
-      .and_return(aws_ssm_client_double)
+      .and_return(aws_ssm_client)
 
-    aws_ssm_client_double
+    aws_ssm_client
+  end
+
+  def stub_main_aws_sts_client
+    aws_sts_client = instance_double(Aws::STS::Client)
+    allow(Aws::STS::Client).to receive(:new).and_return(aws_sts_client)
+    allow(aws_sts_client)
+      .to receive_message_chain(:get_caller_identity, :account)
+      .and_return(CORE_AWS_ACCOUNT_ID)
+    aws_sts_client
   end
 
   private
 
-  def aws_sts_client
+  def stub_aws_sts_client
     aws_sts_client = instance_double(Aws::STS::Client)
     allow(Aws::STS::Client).to receive(:new).and_return(aws_sts_client)
     aws_sts_client