+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - users are shown a warning about the effect of using any action within the service
 - secret environment variables are hidden by default
 - secret environment variables can be shown and hidden individually by clicking the cell
+- infrastructure variables are shown for every infrastructure (not just those running on the core AWS account)
 
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/